feat: Update tooling (#63)
With the imminent change in SASS to switch from @import to @use a more fine grained approach to testing would improve the testing of the changes.

To this end, jest is being added to support that testing.

In addition prettier has been added to the workflow to ensure consistency in the formatting of all the code.